+++ somethingholdingmoney.h
@@ -8,10 +8,10 @@
 class SomethingHoldingMoney
 {
 public:
-    SomethingHoldingMoney(int64_t id, std::string clientLogin, int64_t initialBalance, double percents, time_t creationTime);
+    SomethingHoldingMoney(int64_t id, std::string clientLogin, double initialBalance, double percents, time_t creationTime);
 
     int64_t getId() const { return id; }
-    int64_t getBalance() const { return balance; }
+    double getBalance() const { return balance; }
     double getPercents() const { return percents; }
     time_t getCreationTime() const { return creationTime; }
     std::string getClientLogin() const { return clientLogin; }
@@ -19,7 +19,7 @@ public:
 protected:
     int64_t id;
     std::string clientLogin;
-    int64_t balance;
+    double balance;
     double percents;
     time_t creationTime;
 };
